Bencode and Hash Calculations
Each data structure of the protocol has its own type and the decoder
knows how to create the types. Each type knows how to encode itself
back and also holds the bytes of its values. That is necessary for
because the value of the key 'pieces' is binary.

The hash calculations of the value of the key 'info' as well as
the individual pieces was also implemented.

A full 'torrent' type will be considered to facilitate the way
the key values are extracted in order to make the http calls to
the servers, but this might be implemented only after understanding
what is needed for the call.