Change Pieces constructions to be 0 based index
During the Piece creation the pieces were receiving
its ids starting on 1 instead of 0. That would be a bug
given the piece request messages use piece index 0 based.

It all worked until now because the "download" method was
hardcoding the index as 0, because until this point only the first
piece was been downloaded.